Extract setShouldLogExceptions
